ai-agent/requirements.txt
pymongo

google-genai
pypdf
python-dotenv
requests
pdfminer.six
pytesseract
beautifulsoup4
PyGithub
Pillow
fastapi
uvicorn
python-multipart
pytest
mongomock
httpx
